+++ src/lang/subprogram.hpp
@@ -12,7 +12,9 @@ class Subprogram {
   static int64_t search(const Program &main, const Program &sub,
                         std::map<int64_t, int64_t> &cell_map);
 
-  static bool unfold(Program &main);
+  static bool canUnfold(Operation::Type type);
+
+  static bool unfold(Program &main, int64_t pos = -1);
 
   static bool autoUnfold(Program &main);